+++ src/base64_plantuml.rs
@@ -80,6 +80,19 @@ mod tests {
             String::from("CJ8pD452GqHXOcDa4WW1"),
             Base64PlantUML::encode(&data)
         );
-    }
 
+        //How would one pass 256 here?
+        let data: Vec<u8> = (0 as u8..255 as u8).collect();
+        assert_eq!(
+            String::from(
+                "00420mG51WS82GeB30qE3n0H4XCK5HON61aQ6nmT7XyW8I8Z92Kc9o\
+                 WfAYiiBIulC34oCpGrDZSuEJexF3q-Fq11GaD4HKP7I4bAIqnDJazGKL9JL5LMLr\
+                 XPMbjSNLvVO65YOsHbPcTeQMfhR6rkRt1nSdDqTNPtU7bwUtnzVd-0WOA3X8M6Xu\
+                 Y9YekCZOwFa96IavILbfUOcPgRd9sUdw2XegEafQQdgAcggwojhg-miRApjBMsjx\
+                 YvkhkylRw_mC72myJ5niV8oShBpCtEpz3HqjFKrTRNsDdQszpTtj_WuUBZvENcv-\
+                 ZfwklixUxlyF7oy_JrzlVu-Vhx_Ft-"
+            ),
+            Base64PlantUML::encode(&data)
+        );
+    }
 }