+++ src/plantuml_backend.rs
@@ -6,6 +6,7 @@ use plantuml_server_backend::PlantUMLServer;
 use plantuml_shell_backend::PlantUMLShell;
 use plantumlconfig::PlantUMLConfig;
 use reqwest::Url;
+use uuid::Uuid;
 
 pub trait PlantUMLBackend {
     /// Render a PlantUML string and return the diagram file path (as a String)
@@ -19,6 +20,17 @@ pub fn get_extension(plantuml_code: &String) -> String {
     } else {
         String::from("svg")
     }
+}
+
+/// Create the image names with the appropriate extension and path
+/// The base name of the file is a UUID to avoid collisions with existing
+/// files
+pub fn get_image_filename(img_root: &PathBuf, extension: &String) -> PathBuf {
+    let mut output_file = img_root.clone();
+    output_file.push(Uuid::new_v4().to_string());
+    output_file.set_extension(extension);
+
+    output_file
 }
 
 /// Create an instance of the PlantUMLBackend
@@ -44,5 +56,40 @@ pub fn create(cfg: &PlantUMLConfig, book_root: &PathBuf) -> Box<PlantUMLBackend>
         Box::new(PlantUMLServer::new(server_url, img_root))
     } else {
         Box::new(PlantUMLShell::new(cmd, img_root))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use pretty_assertions::assert_eq;
+
+    #[test]
+    fn test_get_extension() {
+        assert_eq!(
+            String::from("svg"),
+            get_extension(&String::from("C --|> D"))
+        );
+
+        assert_eq!(
+            String::from("png"),
+            get_extension(&String::from("@startditaa"))
+        );
+
+        assert_eq!(
+            String::from("png"),
+            get_extension(&String::from(
+                "Also when not at the start of the code block @startditaa"
+            ))
+        );
+    }
+
+    #[test]
+    fn test_get_image_filename() {
+        let file_path = get_image_filename(&PathBuf::from("foo"), &String::from("bar"));
+
+        assert_eq!(PathBuf::from("foo"), file_path.parent().unwrap());
+        assert_eq!(PathBuf::from("bar"), file_path.extension().unwrap());
+        assert_eq!(36, file_path.file_stem().unwrap().len());
     }
 }